feat(filter-field): Added new methods to abstract data-source class for more flexibility.
To be more flexible when developing custom data-sources or adding new functionallity to the filter-field new abstract methods have been added.

BREAKING CHANGE: Added new methods that have to be implemented when creating a custom data-source
