+++ components/filter-field/src/index.ts
@@ -20,6 +20,7 @@ export * from './filter-field-tag/filter-field-tag';
 export * from './filter-field-range/filter-field-range';
 export * from './filter-field-range/filter-field-range-trigger';
 export * from './filter-field-data-source';
+export * from './filter-field-default-data-source';
 export * from './filter-field-errors';
 
 // @breaking-change 5.0.0 Export only necessary types
